+++ src/modules/blockchain/evm/evm.service.ts
@@ -2,7 +2,7 @@ import { Inject, Injectable, Logger } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { ChainId } from 'src/common/enums';
 import { AppConfig } from 'src/config/app.config';
-import { BigNumberish, ethers, formatUnits, JsonRpcProvider, WebSocketProvider } from 'ethers';
+import { BigNumberish, ethers, formatUnits, JsonRpcProvider, Network, WebSocketProvider } from 'ethers';
 import { ChainIsNotSupportedException } from '../exceptions';
 import { ConfigurationService } from 'src/modules/configuration/configuration.service';
 import { LoggerContext } from 'src/common/decorators/logger-context.decorator';
@@ -38,7 +38,11 @@ export class EvmService {
     if (force || !this.jsonRpcProvidersPool.has(chainId)) {
       this.jsonRpcProvidersPool.get(chainId)?.removeAllListeners();
       const uri = await this.getRPCUrlOnConfigurationOrDefault(chainId);
-      const provider = new JsonRpcProvider(uri);
+      const network = Network.from({
+        chainId: +chainId,
+        name: ChainId[chainId].toLocaleLowerCase(),
+      });
+      const provider = new JsonRpcProvider(uri, network, { staticNetwork: network });
       this.jsonRpcProvidersPool.set(chainId, provider);
       return provider;
     }
@@ -194,9 +198,9 @@ export class EvmService {
       case ChainId.LOCALHOST:
         return `wss://localhost:8545`;
       case ChainId.MOONBEAM:
-        return 'wss://moonbeam.api.onfinality.io/public-ws';
+        return 'wss://moonbeam.unitedbloc.com';
       case ChainId.MOONRIVER:
-        return 'wss://moonriver.api.onfinality.io/public-ws';
+        return 'wss://moonriver.unitedbloc.com';
       case ChainId.MUMBAI:
         return '';
       case ChainId.POLYGON:
@@ -218,9 +222,9 @@ export class EvmService {
       case ChainId.LOCALHOST:
         return `http://localhost:8545`;
       case ChainId.MOONBEAM:
-        return 'https://moonbeam.api.onfinality.io/public';
+        return 'https://moonbeam.unitedbloc.com';
       case ChainId.MOONRIVER:
-        return 'https://moonriver.api.onfinality.io/public';
+        return 'https://moonriver.unitedbloc.com';
       case ChainId.MUMBAI:
         return 'https://rpc-mumbai.maticvigil.com';
       case ChainId.POLYGON: