+++ src/modules/tokens/dtos/contract-address-tokens-request.dto.ts
@@ -28,6 +28,6 @@ export class ContractAddressTokensRequestDto {
   @Transform(({ value }) => (value ? Number(value) : 10))
   @IsNumber()
   @Min(1)
-  @Max(50)
+  @Max(100)
   limit: number = 10;
 }