+++ src/modules/tokens/tokens-jobs-fetch-metadata.service.ts
@@ -11,7 +11,7 @@ import { parallel } from 'radash';
 import { TooManyRequestsExceptionDto } from 'src/common/dtos/http-exception.dto';
 import { RequestHelpers } from 'src/common/helpers/request.helpers';
 import { Optional } from 'src/common/interfaces';
-import { ILike, LessThan, Repository } from 'typeorm';
+import { LessThan, Repository } from 'typeorm';
 import { sanitizeUri } from '../blockchain/evm/utils';
 import { LocalQueueEnum, TokenJobJobs } from '../queue/enums';
 import { TokenAssetEntity } from './entities/tokens-assets.entity';
@@ -212,10 +212,19 @@ export class TokensJobsFetchMetadataService {
     const asset = sanitizePayload.imageRawUrl
       ? await this.assetRepository.findOne({
           where: {
-            rawUrl: ILike(sanitizePayload.imageRawUrl),
+            rawUrl: sanitizePayload.imageRawUrl,
           },
         })
       : null;
+
+    const assetTokenJob =
+      asset?.id && sanitizePayload.imageRawUrl
+        ? await this.tokenJobRepository.findOne({
+            where: {
+              assetUri: sanitizePayload.imageRawUrl,
+            },
+          })
+        : null;
     await this.tokenRepository.update(
       {
         address: params.address,
@@ -231,6 +240,7 @@ export class TokensJobsFetchMetadataService {
         hasMetadata: isNotEmptyObject(sanitizePayload?.metadata),
         assetId: asset?.id,
         hasAsset: isUUID(asset?.id),
+        assetTokenJobId: assetTokenJob?.id,
       },
     );
   }